requirements.txt
annotated-types==0.7.0
APScheduler==3.6.3
asgiref==3.8.1
cachetools==4.2.2
certifi==2024.7.4
charset-normalizer==3.3.2
diff-match-patch==20230430
Django==4.2.8
django-import-export==4.1.1
django-object-actions==4.2.0
et-xmlfile==1.1.0
flake8==7.1.0
gunicorn==23.0.0
idna==3.7
mccabe==0.7.0
openpyxl==3.1.5
prettytable==3.11.0
psycopg2==2.9.9
pycodestyle==2.12.0
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pyflakes==3.2.0
python-dotenv==1.0.1
python-telegram-bot==13.7
pytz==2024.1
requests==2.32.3
schedule==1.2.2
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
vk-api==11.9.9
wcwidth==0.2.13
